+++ Source/Selectivities/Children/InverseLogistic.h
@@ -39,9 +39,9 @@ protected:
 
 private:
   // Members
-  Double                      a50_;
-  Double                      ato95_;
-  Double                      alpha_;
+  float                      a50_;
+  float                      ato95_;
+  float                      alpha_;
 };
 
 } /* namespace selectivities */
